renamed a util I made to a more specific name, updated imports in other files
