+++ client/src/URL.jsx
@@ -1,4 +1,4 @@
-// const URL = "http://localhost:3001"
-const URL = "https://mindful-journal-server.vercel.app"
+const URL = "http://localhost:3001"
+// const URL = "https://mindful-journal-server.vercel.app"
 
 export default URL;
